+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- None
+### Added
+
+- Support to persist rich text editor layout ([#4](https://github.com/benji300/joplin-persistent-layout/issues/4))
 
 ## [1.0.0] - 2021-01-29